+++ src/App.jsx
@@ -1,8 +1,11 @@
 import React from 'react';
+import Login from './components/Login';
 
 function App() {
   return (
-    <div className="App" />
+    <div className="App">
+      <Login />
+    </div>
   );
 }
